Moved the options menu from the main Activity to the movieListFragment
